making config.json work with the replicator
