+++ config.json
@@ -9,7 +9,8 @@
   "log_level": "info",
   "own_registry_adr": "127.0.0.1",
   "own_registry_port": "8585",
-  "states": ["https://registry.bupd.xyz/satellite/group10/state:latest"],
+  "states": ["https://registry.bupd.xyz/satellite-test-group-state/state:latest"],
   "url_or_file": "https://registry.bupd.xyz",
-  "zotconfigpath": "./registry/config.json"
+  "zotconfigpath": "./registry/config.json",
+  "use_unsecure": true
 }
