+++ internal/state/state_process.go
@@ -6,6 +6,7 @@ import (
 	"fmt"
 	"sync"
 
+	"container-registry.com/harbor-satellite/internal/config"
 	"container-registry.com/harbor-satellite/internal/notifier"
 	"container-registry.com/harbor-satellite/internal/utils"
 	"container-registry.com/harbor-satellite/logger"
@@ -25,26 +26,38 @@ type FetchAndReplicateAuthConfig struct {
 }
 
 type FetchAndReplicateStateProcess struct {
-	id                   uint64
-	name                 string
-	stateArtifactFetcher StateFetcher
-	cronExpr             string
-	isRunning            bool
-	stateReader          StateReader
-	notifier             notifier.Notifier
-	mu                   *sync.Mutex
-	authConfig           FetchAndReplicateAuthConfig
+	id         uint64
+	name       string
+	cronExpr   string
+	isRunning  bool
+	stateMap   []StateMap
+	notifier   notifier.Notifier
+	mu         *sync.Mutex
+	authConfig FetchAndReplicateAuthConfig
 }
 
-func NewFetchAndReplicateStateProcess(id uint64, cronExpr string, stateFetcher StateFetcher, notifier notifier.Notifier, username, password, remoteRegistryURL, sourceRegistryURL string, useUnsecure bool) *FetchAndReplicateStateProcess {
+type StateMap struct {
+	url   string
+	State StateReader
+}
+
+func NewStateMap(url []string) []StateMap {
+	var stateMap []StateMap
+	for _, u := range url {
+		stateMap = append(stateMap, StateMap{url: u, State: nil})
+	}
+	return stateMap
+}
+
+func NewFetchAndReplicateStateProcess(id uint64, cronExpr string, notifier notifier.Notifier, username, password, remoteRegistryURL, sourceRegistryURL string, useUnsecure bool, states []string) *FetchAndReplicateStateProcess {
 	return &FetchAndReplicateStateProcess{
-		id:                   id,
-		name:                 FetchAndReplicateStateProcessName,
-		cronExpr:             cronExpr,
-		isRunning:            false,
-		stateArtifactFetcher: stateFetcher,
-		notifier:             notifier,
-		mu:                   &sync.Mutex{},
+		id:        id,
+		name:      FetchAndReplicateStateProcessName,
+		cronExpr:  cronExpr,
+		isRunning: false,
+		notifier:  notifier,
+		mu:        &sync.Mutex{},
+		stateMap:  NewStateMap(states),
 		authConfig: FetchAndReplicateAuthConfig{
 			Username:          username,
 			Password:          password,
@@ -62,32 +75,49 @@ func (f *FetchAndReplicateStateProcess) Execute(ctx context.Context) error {
 		return fmt.Errorf("process %s already running", f.GetName())
 	}
 	defer f.stop()
-	newStateFetched, err := f.FetchAndProcessState(log)
-	if err != nil {
-		return err
+
+	for _, state := range f.stateMap {
+		log.Info().Msgf("Processing state for %s", state.url)
+		stateFetcher, err := processInput(state.url, f.authConfig.Username, f.authConfig.Password, log)
+		if err != nil {
+			log.Error().Err(err).Msg("Error processing input")
+			return err
+		}
+		newStateFetched, err := f.FetchAndProcessState(stateFetcher, log)
+		if err != nil {
+			log.Error().Err(err).Msg("Error fetching state")
+			return err
+		}
+		log.Info().Msgf("State fetched successfully for %s", state.url)
+		deleteEntity, replicateEntity, newState := f.GetChanges(newStateFetched, log, state.State)
+		f.LogChanges(deleteEntity, replicateEntity, log)
+		if err := f.notifier.Notify(); err != nil {
+			log.Error().Err(err).Msg("Error sending notification")
+		}
+		log.Info().Msg("Replicating state")
+		log.Info().Msg("Replicator config")
+		log.Info().Msgf("Username: %s", f.authConfig.Username)
+		log.Info().Msgf("Password: %s", f.authConfig.Password)
+		log.Info().Msgf("Remote registry URL: %s", f.authConfig.remoteRegistryURL)
+		log.Info().Msgf("Source registry: %s", f.authConfig.sourceRegistry)
+		
+		replicator := NewBasicReplicator(f.authConfig.Username, f.authConfig.Password, f.authConfig.remoteRegistryURL, f.authConfig.sourceRegistry, f.authConfig.useUnsecure)
+		// Delete the entities from the remote registry
+		if err := replicator.DeleteReplicationEntity(ctx, deleteEntity); err != nil {
+			log.Error().Err(err).Msg("Error deleting entities")
+			return err
+		}
+		// Replicate the entities to the remote registry
+		if err := replicator.Replicate(ctx, replicateEntity); err != nil {
+			log.Error().Err(err).Msg("Error replicating state")
+			return err
+		}
+		state.State = newState
 	}
-	log.Info().Msg("State fetched successfully")
-	deleteEntity, replicateEntity, newState := f.GetChanges(newStateFetched, log)
-	f.LogChanges(deleteEntity, replicateEntity, log)
-	if err := f.notifier.Notify(); err != nil {
-		log.Error().Err(err).Msg("Error sending notification")
-	}
-	replicator := NewBasicReplicator(f.authConfig.Username, f.authConfig.Password, f.authConfig.remoteRegistryURL, f.authConfig.sourceRegistry, f.authConfig.useUnsecure)
-	// Delete the entities from the remote registry
-	if err := replicator.DeleteReplicationEntity(ctx, deleteEntity); err != nil {
-		log.Error().Err(err).Msg("Error deleting entities")
-		return err
-	}
-	// Replicate the entities to the remote registry
-	if err := replicator.Replicate(ctx, replicateEntity); err != nil {
-		log.Error().Err(err).Msg("Error replicating state")
-		return err
-	}
-	f.stateReader = newState
 	return nil
 }
 
-func (f *FetchAndReplicateStateProcess) GetChanges(newState StateReader, log *zerolog.Logger) ([]ArtifactReader, []ArtifactReader, StateReader) {
+func (f *FetchAndReplicateStateProcess) GetChanges(newState StateReader, log *zerolog.Logger, oldState StateReader) ([]ArtifactReader, []ArtifactReader, StateReader) {
 	log.Info().Msg("Getting changes")
 	// Remove artifacts with null tags from the new state
 	newState = f.RemoveNullTagArtifacts(newState)
@@ -95,14 +125,14 @@ func (f *FetchAndReplicateStateProcess) GetChanges(newState StateReader, log *ze
 	var entityToDelete []ArtifactReader
 	var entityToReplicate []ArtifactReader
 
-	if f.stateReader == nil {
+	if oldState == nil {
 		log.Warn().Msg("Old state is nil")
 		return entityToDelete, newState.GetArtifacts(), newState
 	}
-	
+
 	// Create maps for quick lookups
 	oldArtifactsMap := make(map[string]ArtifactReader)
-	for _, oldArtifact := range f.stateReader.GetArtifacts() {
+	for _, oldArtifact := range oldState.GetArtifacts() {
 		tag := oldArtifact.GetTags()[0]
 		oldArtifactsMap[oldArtifact.GetName()+"|"+tag] = oldArtifact
 	}
@@ -199,9 +229,9 @@ func ProcessState(state *StateReader) (*StateReader, error) {
 	return state, nil
 }
 
-func (f *FetchAndReplicateStateProcess) FetchAndProcessState(log *zerolog.Logger) (StateReader, error) {
+func (f *FetchAndReplicateStateProcess) FetchAndProcessState(fetcher StateFetcher, log *zerolog.Logger) (StateReader, error) {
 	state := NewState()
-	err := f.stateArtifactFetcher.FetchStateArtifact(&state)
+	err := fetcher.FetchStateArtifact(&state)
 	if err != nil {
 		log.Error().Err(err).Msg("Error fetching state artifact")
 		return nil, err
@@ -213,4 +243,45 @@ func (f *FetchAndReplicateStateProcess) FetchAndProcessState(log *zerolog.Logger
 func (f *FetchAndReplicateStateProcess) LogChanges(deleteEntity, replicateEntity []ArtifactReader, log *zerolog.Logger) {
 	log.Warn().Msgf("Total artifacts to delete: %d", len(deleteEntity))
 	log.Warn().Msgf("Total artifacts to replicate: %d", len(replicateEntity))
+}
+
+func processInput(input, username, password string, log *zerolog.Logger) (StateFetcher, error) {
+
+	if utils.IsValidURL(input) {
+		return processURLInput(utils.FormatRegistryURL(input), username, password, log)
+	}
+
+	log.Info().Msg("Input is not a valid URL, checking if it is a file path")
+	if err := validateFilePath(input, log); err != nil {
+		return nil, err
+	}
+
+	return processFileInput(input, username, password, log)
+}
+
+func validateFilePath(path string, log *zerolog.Logger) error {
+	if utils.HasInvalidPathChars(path) {
+		log.Error().Msg("Path contains invalid characters")
+		return fmt.Errorf("invalid file path: %s", path)
+	}
+	if err := utils.GetAbsFilePath(path); err != nil {
+		log.Error().Err(err).Msg("No file found")
+		return fmt.Errorf("no file found: %s", path)
+	}
+	return nil
+}
+
+func processURLInput(input, username, password string, log *zerolog.Logger) (StateFetcher, error) {
+	log.Info().Msg("Input is a valid URL")
+	config.SetRemoteRegistryURL(input)
+
+	stateArtifactFetcher := NewURLStateFetcher(input, username, password)
+
+	return stateArtifactFetcher, nil
+}
+
+func processFileInput(input, username, password string, log *zerolog.Logger) (StateFetcher, error) {
+	log.Info().Msg("Input is a valid file path")
+	stateArtifactFetcher := NewFileStateFetcher(input, username, password)
+	return stateArtifactFetcher, nil
 }
